requirements.txt
typing_extensions==4.15.0
winrt-runtime==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Media.Control==3.2.1
